Fix IllegalArgumentException for ping event?
Not 100% confident that this will fix it.
Please **do tests before merging this**!

I essentially replaced the `createProfile` method with the one requiring a UUID and use the `UUID.randomUUID()` to handle this one.
Hopefully this will fix the issue I mentioned in #1

Also, note that as stated in https://github.com/AlexProgrammerDE/PistonMOTD/issues/1#issuecomment-731644584 does the playercounter still get modified despite having the "activated" method disabled. This should be looked into.