sponge: Consider all metrics collection states
SpongeAPI has 3 potential collection states that a plugin can
effectively have:
1. TRUE -> The server administrator has explicitly enabled metrics for
   either the plugin, or the server globally.
2. FALSE -> The server administrator has explicitly disabled metrics
   for either the plugin, or the server globally.
3. UNDEFINED -> The server administrator has set neither a plugin or
   global collection state. This defaults to metrics being disabled.

If a server administrator has explicitly or implicitly disabled
metrics globally, it is still possible that metrics may be enabled for
a plugin - through its own collection state, which has a higher
precedence when establishing the effective collection state.

This also works conversely - its entirely possible that a global state
could be TRUE, but a plugin's effective state is still FALSE - if the
server administrator has explicitly disabled metrics for that plugin.